+++ client/src/pages/Login.js
@@ -166,46 +166,6 @@ const Login = () => {
     password: '',
   })
 
-  const url = new URL(window.location.href)
-  const authorizationCode = url.searchParams.get('code')
-  const stateCode = url.searchParams.get('state')
-
-  //받은 authorization 코드이용 서버로 callback api 요청
-  const getAccessTocken = async authorizationCode => {
-    await axios
-      .post(
-        'http://localhost:8081/snslogin/gettoken',
-        { authorizationCode: authorizationCode, stateCode: stateCode },
-        {
-          withCredentials: true,
-        }
-      )
-      .then(res => {
-        getSnsInfo()
-      })
-      .catch(err => {
-        alert('서버로부터 응답이 없습니다. 나중에 다시 시도해주세요')
-      })
-  }
-
-  const getSnsInfo = async () => {
-    await axios
-      .get('http://localhost:8081/snslogin/getuserinfo', { withCredentials: true })
-      .then(res => {
-        dispatch(changeToLoginTrue())
-        dispatch(changeToDiaryTrue())
-        dispatch(welcomeMode())
-        navigate('/')
-      })
-      .catch(err => {
-        alert('서버로부터 응답이 없습니다. 나중에 다시 시도해주세요')
-      })
-  }
-
-  useEffect(() => {
-    // getAccessTocken(authorizationCode)
-  }, [])
-
   const handleInputValue = key => e => {
     setLoginInfo({ ...loginInfo, [key]: e.target.value })
   }
