+++ client/src/pages/SnsLoading.css
@@ -1,0 +1,64 @@
+/* .loadingBox .circle {
+  position: fixed;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  width: 40px;
+  height: 40px;
+  border: 10px solid #fff;
+  border-top: 10px solid rgb(94, 93, 93);
+  border-radius: 50em;
+  animation-name: spinCircle;
+  animation-duration: 0.8s;
+  animation-iteration-count: infinite;
+}
+
+@keyframes spinCircle {
+  from {
+    transform: translate(-50%, -50%) rotate(0);
+  }
+  to {
+    transform: translate(-50%, -50%) rotate(360deg);
+  }
+} */
+.loading-box {
+  position: fixed;
+  left: 50%;
+  top: 50%;
+  z-index: 100;
+  transform: translate(-50%, -50%);
+  padding: 20px;
+  width: 200px;
+  text-align: center;
+  background: #fff;
+}
+.loading-box .circles {
+  padding-top: 10px;
+}
+.loading-box .circles i {
+  animation: scaleBounce 0.3s alternate infinite;
+  display: inline-block;
+  margin: 0 4px;
+  width: 10px;
+  height: 10px;
+  background: #404242;
+  border-radius: 50em;
+}
+.loading-box .circles i:nth-child(2) {
+  animation-delay: 0.1s;
+}
+.loading-box .circles i:nth-child(3) {
+  animation-delay: 0.2s;
+}
+.loading-box p {
+  margin-top: 20px;
+  font-size: 18px;
+}
+@keyframes scaleBounce {
+  from {
+    transform: scale(0.7);
+  }
+  to {
+    transform: scale(1.3);
+  }
+}
