+++ server/src/routes/snslogin/post/index.ts
@@ -33,7 +33,7 @@ export default {
       const params = new URLSearchParams()
       params.append('grant_type', 'authorization_code')
       params.append('client_id', '7a15a8d44b88c4a6cc057ca28ad75307')
-      params.append('redirect_uri', 'http://localhost:3000/login')
+      params.append('redirect_uri', 'http://localhost:3000/sns')
       params.append('code', req.body.authorizationCode)
 
       await axios
